Add (limited) support for spitting transactions
The YNAB APIs don't allow you to split a tranasaction that already
exists. To work around this for an imported transaction the
`prepare_split_for_matching` tool will recreate the transaction that
exists with the same details as the original but use sub transactions to
split it. From there you can go into the YNAB app and match the
transaction.